+++ tests-corpus/valid-bare-schema/enum-unsorted/out.gen.d.ts
@@ -1,0 +1,11 @@
+import * as bare from "@bare-ts/lib"
+
+export declare enum Gender {
+    FLUID = "FLUID",
+    MALE = "MALE",
+    FEMALE = "FEMALE",
+}
+
+export function readGender(bc: bare.ByteCursor): Gender
+
+export function writeGender(bc: bare.ByteCursor, x: Gender): void